requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-bootstrap-form==3.4
django-simple-history==3.1.1
dnspython==2.0.0
ipwhois==1.2.0
psycopg2-binary==2.9.3
sqlparse==0.4.2
requests